1.auto font config; 2.argument to specify whether to download and compile from source

+++ setup_vim8.sh
@@ -1,33 +1,36 @@
-# install ncurses
-cd ~/repo
-wget ftp.gnu.org/pub/gnu/ncurses/ncurses-6.1.tar.gz
-tar -xzf ncurses-6.1.tar.gz
-cd ncurses-6.1
-./configure --prefix=/home/yetiancai/.local/
-make
-make install
+if [[ $# -eq 1 && $1 == "--compile" ]]; then
+    # install ncurses
+    cd ~/repo
+    wget ftp.gnu.org/pub/gnu/ncurses/ncurses-6.1.tar.gz
+    tar -xzf ncurses-6.1.tar.gz
+    cd ncurses-6.1
+    ./configure --prefix=/home/yetiancai/.local/
+    make
+    make install
 
-cd ..
-rm ncurses-6.1* -rf
+    cd ..
+    rm ncurses-6.1* -rf
 
-# install vim8
-git clone https://github.com/vim/vim.git
-cd vim
-git checkout v8.1.0264
-./configure --prefix=/home/yetiancai/.local/ \
-    --with-features=huge \
-    --enable-multibyte \
-    --enable-rubyinterp=dynamic \
-    --enable-pythoninterp=dynamic \
-    --enable-python3interp=dynamic \
-    --enable-perlinterp=dynamic \
-    --enable-luainterp=dynamic \
-    --enable-gui=no \
-    --enable-cscope \
-    --enable-terminal
-make
-make install
+    # install vim8
+    git clone https://github.com/vim/vim.git
+    cd vim
+    git checkout v8.1.0264
+    ./configure --prefix=/home/yetiancai/.local/ \
+        --with-features=huge \
+        --enable-multibyte \
+        --enable-rubyinterp=dynamic \
+        --enable-pythoninterp=dynamic \
+        --enable-python3interp=dynamic \
+        --enable-perlinterp=dynamic \
+        --enable-luainterp=dynamic \
+        --enable-gui=no \
+        --enable-cscope \
+        --enable-terminal
+    make
+    make install
+fi
 
 echo source ~/repo/editor_config/vimrc >> ~/.vimrc
+cp ~/repo/editor_config/fonts ~/.fonts -r
 curl -fLo ~/.vim/autoload/plug.vim --create-dirs \
     https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
